init wayland support.
The display server based on X or Wayland, is now a choice of different profi
Basically and consistently, two macros were used "with wayland" and "with x"
Below summarize the combination of the macros:

| wayland |  x  | meaning
|---------------------------
|   0     |  1  | pure X11 platform(no wayland)
|   1     |  0  | pure wayland platform (no X11)
|   1     |  1  | wayland but X compatibility
|   0     |  0  | no X and no wayland

This method unifies the meaning and usage. Deploy this method to menu-screen package.
Currently we need "pure wayland platform (no X11)" status,so "with wayland"
defined into "1" and "0" in build.conf file.If you need other status of abov
please change values of "with x" and "with wayland" in your build.conf file.

Signed-off-by: wangfei <feix.w.wang@intel.com>
Change-Id: I9b79da3c5710a559567c06b74d04d2e593d76a35